pyproject.toml
[project]
name = "django-saas-sso"
version = "0.1.0"
description = "The domain module for building a SaaS product with Django"
authors = [
    { name = "Hsiaoming Yang", email = "me@lepture.com" }
]
dependencies = [
    "Django",
    "djangorestframework",
    "requests",
    "joserfc",
    "django-saas-base",
]
readme = "README.md"
requires-python = ">= 3.10"
license = {text = "BSD-3-Clause"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "drf_spectacular",
    "pytest",
    "pytest-django",
    "pytest-cov",
    "requests-mock",
]

[tool.rye.scripts]
test-cov = "pytest --cov --cov-report=html"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/saas_domain"]

[tool.hatch.build.targets.sdist]
exclude = [
    ".python-version",
    ".github/*",
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
addopts = "--no-migrations"
pythonpath = ["src"]
testpaths = ["tests"]
filterwarnings = ["error"]
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["saas_domain"]
omit = ["*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "@(abc\\.)?abstractmethod",
]
